+++ packages/syntax-core/src/Avatar/Avatar.tsx
@@ -21,7 +21,10 @@ const Avatar = ({
   /**
    * Size of the avatar.
    *
-   * `sm`: 24px, `md`: 40px, `lg`: 72px, `xl`: 128px
+   * * `sm`: 24px
+   * * `md`: 40px
+   * * `lg`: 72px
+   * * `xl`: 128px
    *
    * @defaultValue `md`
    */