+++ packages/syntax-core/src/Button/Button.tsx
@@ -59,7 +59,9 @@ const Button = ({
   /**
    * The size of the button
    *
-   * `sm`: 32px, `md`: 40px, `lg`: 48px
+   * * `sm`: 32px
+   * * `md`: 40px
+   * * `lg`: 48px
    *
    * @defaultValue "md"
    */