+++ apps/storybook/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @syntax/storybook
 
+## 0.6.2
+
+### Patch Changes
+
+- Updated dependencies [3309cdb]
+  - @cambly/syntax-core@1.14.0
+
+## 0.6.1
+
+### Patch Changes
+
+- Updated dependencies [6c1dd9b]
+  - @cambly/syntax-core@1.13.0
+
 ## 0.6.0
 
 ### Minor Changes