+++ packages/syntax-core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @cambly/syntax-core
 
+## 1.14.0
+
+### Minor Changes
+
+- 3309cdb: Properly export Heading and Typography
+
+## 1.13.0
+
+### Minor Changes
+
+- 6c1dd9b: Add Typography and Heading components
+
 ## 1.12.0
 
 ### Minor Changes